uint8 -> uint256

+++ source/contracts/factories/ShareTokenFactory.sol
@@ -8,7 +8,7 @@ import 'IController.sol';
 
 
 contract ShareTokenFactory {
-    function createShareToken(IController _controller, IMarket _market, uint8 _outcome) public returns (IShareToken) {
+    function createShareToken(IController _controller, IMarket _market, uint256 _outcome) public returns (IShareToken) {
         Delegator _delegator = new Delegator(_controller, "ShareToken");
         IShareToken _shareToken = IShareToken(_delegator);
         _shareToken.initialize(_market, _outcome);
