+++ source/contracts/reporting/IReportingParticipant.sol
@@ -13,7 +13,7 @@ contract IReportingParticipant {
     function isInvalid() public view returns (bool);
     function isDisavowed() public view returns (bool);
     function migrate() public returns (bool);
-    function getPayoutNumerator(uint8 _outcome) public view returns (uint256);
+    function getPayoutNumerator(uint256 _outcome) public view returns (uint256);
     function getMarket() public view returns (IMarket);
     function getSize() public view returns (uint256);
 }
