+++ tests/solidity_test_helpers/MockFeeWindow.sol
@@ -208,7 +208,7 @@ contract MockFeeWindow is Initializable, MockVariableSupplyToken, IFeeWindow {
         return true;
     }
 
-    function createMarket(uint256 _endTime, uint256 _feePerEthInWei, ICash _denominationToken, address _designatedReporterAddress, address _sender, uint8 _numOutcomes, uint256 _numTicks) public payable returns (IMarket _newMarket) {
+    function createMarket(uint256 _endTime, uint256 _feePerEthInWei, ICash _denominationToken, address _designatedReporterAddress, address _sender, uint256 _numOutcomes, uint256 _numTicks) public payable returns (IMarket _newMarket) {
         return market;
     }
