+++ tests/solidity_test_helpers/MockInitialReporter.sol
@@ -90,7 +90,7 @@ contract MockInitialReporter is IInitialReporter {
         return true;
     }
 
-    function getPayoutNumerator(uint8 _outcome) public view returns (uint256) {
+    function getPayoutNumerator(uint256 _outcome) public view returns (uint256) {
         return 0;
     }
